requirements.txt
aiogram>=3.0.0
aiohttp
python-dotenv
numpy
matplotlib
rauth
fatsecret
